Adding ability to update expenses; beginning of budget pages

+++ fintrack/fintrack.rb
@@ -15,6 +15,10 @@ require_relative 'common.rb'
 
 PATH_ROOT = '/'
 
+LOCATION_FILTER_REGEX = /[^a-zA-Z0-9\/,\. #\$\*_\+!]/
+# do not allow commas in tags since we split on commas
+TAG_FILTER_REGEX = /[^a-zA-Z0-9\/\. #\$\*_\+!]/
+
 class Fintrack < Sinatra::Base
 
   def agg_locals(overrides = {})
@@ -23,6 +27,13 @@ class Fintrack < Sinatra::Base
         'home' => PATH_ROOT,
         'expense_post' => PATH_ROOT + 'expense',
         'expense_delete' => PATH_ROOT + 'expense/delete/%id%',
+        'expense_edit' => PATH_ROOT + 'expense/edit/%id%',
+        'expense_update' => PATH_ROOT + 'expense/update',
+
+        'budget_delete' => PATH_ROOT + 'budget/delete/%id%',
+        'budget_post' => PATH_ROOT + 'budget',
+        'budget_edit' => PATH_ROOT + 'budget/edit/%id%',
+        'budget_update' => PATH_ROOT + 'budget/update',
       }
     }
 
@@ -45,7 +56,7 @@ class Fintrack < Sinatra::Base
     puts "Location starting with #{location}"
     location = location.strip.gsub(/(['])/, '\'\1')
     puts "Location now #{location}"
-    location = location.gsub(/[^a-zA-Z0-9\/\\',\. #\$\*_\+!]/, '')
+    location = location.gsub(LOCATION_FILTER_REGEX, '')
     puts "Location finally #{location}"
     result = execute_query "SELECT Id FROM #{LOCATION_TABLE} WHERE Name='#{location}';"
 
@@ -62,7 +73,7 @@ class Fintrack < Sinatra::Base
   def tag_id(tag)
     id = nil
     tag = tag.strip.gsub(/(['])/, '\'\1')
-    tag = tag.gsub(/[^a-zA-Z0-9\/\\',\. #\$\*_\+!]/, '')
+    tag = tag.gsub(TAG_FILTER_REGEX, '')
     result = execute_query "SELECT Id FROM #{TAG_TABLE} WHERE Name='#{tag.downcase}';"
 
     if result.size == 1
@@ -104,12 +115,14 @@ class Fintrack < Sinatra::Base
   def expenses(opts = {})
     start_date = opts['start_date'] || beginning_of_month
     end_date = opts['end_date'] || end_of_month
+    ids = opts['ids']
 
     result = execute_query <<-SQL
       SELECT
         Id, Date, Amount, Location, Spender, Entered, Updated, Notes
       FROM #{EXPENSE_TABLE}
       WHERE Date >= #{start_date} and Date <= #{end_date}
+      #{ids ? "and Id IN (#{ids.join(',')})" : ''}
       ORDER BY Date ASC;
     SQL
 
@@ -128,10 +141,14 @@ class Fintrack < Sinatra::Base
       }
 
       map['date'] = "#{map['date'][0...4]}-#{map['date'][4...6]}-#{map['date'][6...8]}"
-      map['amount'] = map['amount'].to_f / 100.0
+      map['amount'] = sprintf("%0.2f", map['amount'].to_f / 100.0)
 
       map
     end
+  end
+
+  def expense_for(id)
+    expenses('start_date' => MIN_DATE, 'end_date' => MAX_DATE, 'ids' => [id])[0]
   end
 
   #
@@ -185,27 +202,52 @@ class Fintrack < Sinatra::Base
     location_id(location).to_s
   end
 
-  get '/ajax/expenses/add' do
-    vals = ['20140103000000', 1803, "This was a trip to 'Publix'"]
-    res = insert(EXPENSE_TABLE, %w(Date Amount Notes), vals)
-    res.to_s
-  end
-
   get '/ajax/expenses' do
     expenses(params).to_json
   end
 
-  get '/ajax/dbdate' do
-    date = execute_query "SELECT date('now');"
-    date.to_json
+  get '/ajax/expense/:id' do |id|
+    expense_for(id).to_json
   end
 
+  #
+  # GET/POST Helpers (to reduce code duplication)
+  #
+
+  def budget_edit(params = {})
+    bom = beginning_of_month
+    lm = MAX_DATE
+
+    period_days = params['period_days'] || '30'
+    start_date = params['start_date'] || "#{bom[0...4]}-#{bom[4...6]}-#{bom[6...8]}"
+    end_date = params['end_date'] || "#{lm[0...4]}-#{lm[4...6]}-#{lm[6...8]}"
+    #end_date = params['end_date'] ? params['end_date'].gsub('-', '') : beginning_of_month
+    name = params['name'] || ''
+    amount = params['amount'] || '0.00'
+    tag = params['tag'] || ''
+
+    budget = {
+      'period_days' => period_days,
+      'start_date' => start_date,
+      'end_date' => end_date,
+      'name' => name,
+      'amount' => amount,
+      'tag' => tag,
+    }
+
+    page_vars = {
+      'budget' => budget,
+    }
+
+    erb :budget_edit, :locals => agg_locals(page_vars)
+  end
 
   #
   # POST
   #
 
-  post '/expense' do
+  def expense_edit(params = {})
+    id = (params['id'] || '0').to_i
     date = params['date'].gsub('-', '') + '000000'
     amount = (params['amount'].to_f * 100).round.to_i
     location = params['location']
@@ -240,9 +282,52 @@ class Fintrack < Sinatra::Base
     redirect to(PATH_ROOT)
   end
 
-  get '/expense/delete/:id' do |id|
-    res = execute_query("DELETE FROM #{EXPENSE_TABLE} WHERE Id=#{id.to_i};")
-    res = execute_query("DELETE FROM #{EXPENSE_TAG_TABLE} WHERE Expense=#{id.to_i};")
+  post '/expense' do
+    expense_edit(params)
+  end
+
+  post '/expense/update' do
+    id = params['id'].to_i
+    date = params['date'].gsub('-', '') + '000000'
+    amount = (params['amount'].to_f * 100).round.to_i
+    location = params['location']
+    tags = params['tags'].split(',')
+    notes = params['notes']
+    updated = current_datetime
+
+    puts "params: #{params.inspect}"
+
+    if id > 0 and !params['confirm_delete']
+      loc_id = location_id(location)
+
+      vals = [id, date, amount, loc_id, notes, updated]
+      puts "vals: #{vals}"
+
+      res = update(EXPENSE_TABLE, %w(Id Date Amount Location Notes Updated), vals, id)
+
+      if res > 0
+        # delete all current tag associations
+        execute_query("DELETE FROM #{EXPENSE_TAG_TABLE} WHERE Expense=#{id.to_i};")
+
+        tags = tags.reject {|tag| tag.strip.length == 0}
+        tag_ids = tags.map do |tag|
+          tag_id(tag)
+        end
+        tag_ids.reject! {|i| !i || i <= 0}
+
+        # apply new associations
+        tag_ids.each do |tag_id|
+          insert(EXPENSE_TAG_TABLE, %w(Tag Expense), [tag_id, id])
+        end
+
+        puts "Added expense ID #{res}"
+      end
+      redirect to("#{PATH_ROOT}expense/edit/#{id}?updated=true")
+    elsif id > 0 and params['confirm_delete']
+      puts 'This should delete the item!'
+      res = execute_query("DELETE FROM #{EXPENSE_TABLE} WHERE Id=#{id.to_i};")
+      res = execute_query("DELETE FROM #{EXPENSE_TAG_TABLE} WHERE Expense=#{id.to_i};")
+    end
 
     redirect to(PATH_ROOT)
   end
@@ -263,4 +348,26 @@ class Fintrack < Sinatra::Base
 
     erb :expenses, :locals => agg_locals(page_vars)
   end
+
+  get '/budget/add' do
+    budget_edit
+  end
+
+  get '/expense/edit/:id' do |id|
+    page_vars = {
+      'expense' => expense_for(id),
+      'updated' => params['updated'] || false,
+    }
+
+    erb :expense_edit, :locals => agg_locals(page_vars)
+  end
+
+  get '/expense/delete/:id' do |id|
+    res = execute_query("DELETE FROM #{EXPENSE_TABLE} WHERE Id=#{id.to_i};")
+    res = execute_query("DELETE FROM #{EXPENSE_TAG_TABLE} WHERE Expense=#{id.to_i};")
+
+    redirect to(PATH_ROOT)
+  end
+
+
 end
